requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
community==1.0.0b1
contourpy==1.1.1
cycler==0.12.1
dash==2.16.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_color_picker==0.0.1
dnspython==2.6.1
Flask==3.0.0
fonttools==4.43.1
idna==3.4
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mysql-connector-python==8.3.0
neo4j==5.19.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
plotly==5.20.0
pygraphviz==1.11
pymongo==4.6.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-ipd==1.4.0
python-louvain==0.16
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
scipy==1.11.3
six==1.16.0
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==3.0.1
zipp==3.18.1
